+++ public/js/modules/subtitle/export/exportHandlers.js
@@ -1,12 +1,10 @@
-import storageModule from '../../utils/storage.js';
+import { storage } from '../../utils/storage.js';
 import { bindEvent } from '../../utils/dom.js';
 import { encodeText } from './download/iconv.js';
 import { formatExportData } from './formatExportData.js';
 import { downloadBytes } from './download/downloadBytes.js';
 import { buildSmi, buildSrt, buildVtt, buildJson } from './build/buildFile.js';
 import { buildExcel } from './build/buildExcel.js';
-
-const storage = storageModule();
 
 const setStorageAndField = (storageKey, fieldSelector) => (value) => {
 	storage.set(storageKey, value);