+++ public/js/modules/utils/storage.js
@@ -1,17 +1,15 @@
-const storage = () => {
-	const get = (key) => JSON.parse(localStorage.getItem(key));
+const get = (key) => JSON.parse(localStorage.getItem(key));
 
-	const set = (key, value) => {
-		localStorage.setItem(key, JSON.stringify(value));
-		return true;
-	};
-
-	const remove = (key) => {
-		localStorage.removeItem(key);
-		return true;
-	};
-
-	return { get, set, remove };
+const set = (key, value) => {
+	localStorage.setItem(key, JSON.stringify(value));
+	return true;
 };
 
-export default storage;
+const remove = (key) => {
+	localStorage.removeItem(key);
+	return true;
+};
+
+const storage = { get, set, remove };
+
+export { get, set, remove, storage };